fix: Sqitch `deploy` in the container entrypoint when using SQLite.
It's safe to do this, as we assume that the container orchestrator
will only mount the SQLite volume on a single container instance, and
therefore we don't need to worry about atomicity in the deploy
scripts. The Docker container NixOS tests have been changed to reflect
this.

(Note that when running against a PostgreSQL database, the behavior of
the entrypoint is the same as before: it only runs `sqitch verify`,
and assumes that any deployment step has already been done by the time
the `primer-service` container has launched, as this is the only safe
way to do it in a multi-container environment: see
https://github.com/sqitchers/sqitch/discussions/646)