refactor: use simultaneous substitution where applicable
Everywhere we substitute multiple types we actually do not care whether
the substitution is iterated or simultaneous. The only non-singleton
substitutions we use are from one of
- 'genInstApp's output (or the accumulated substitution in its helper 'reify')
- 'unify's output
These are both idempotent (i.e. the substituted types do not refer to
any of the names being substituted away), and thus both sorts of
substitution will return the same value. Thus we should use the more
efficient simultaneous substitution (which only has to traverse the tree
once).

Now that 'substTyIter' is unused, we temporarily tell Weeder to ignore
it. We will remove the dead code in a separate commit, so it is easy to
revive in the future, if necessary.