feat!: ConstructSaturatedCon action creates a saturated constructor
(In the sense of "uses `conSat`" / "embeds the arguments in the `Expr`
subtree rooted at a `Cons` node". Before this commit it created a
constructor applied to these same arguments with `APP` and `App`.)

Note this only changes `ConstructSaturatedCon`, not
`ConstructRefinedCon`, and thus in an openapi frontend using the
`MakeConSat` action the behaviour (i.e. what underlying AST is
created) will differ depending whether it is in a position with no
type info (i.e. where the type is a hole). In particular, in the
program
  Cons @[?_1] [?_2, ?_3]
we have
- ?_2 : ?_1, so we would do a conSat here
- ?_3 : List ?_1, but we would not here (do a refine)

This is a step towards all constructors being saturated, at which
point we can update the typechecker to enforce that invariant.

BREAKING CHANGE: this will change the interpretation of some actions
that may be stored in logs inside programs created from older primer
versions. Thus it may break usage of older databases. (The programs
should still load fine, and be editable, but undo may be broken if a
ConstructSaturatedCon action happens to be in the log.)