constructSatCon action uses sat con...
Now it actually creates a conSat. However, this only happens if we use
the ConstructSaturatedCon, not ConstructRefinedCon, and thus in an
openapi frontend only happens using the MakeConSat action in a position
with no type info (i.e. where the type is a hole). In particular, in the
program
  Cons @[?_1] [?_2, ?_3]
we have
- ?_2 : ?_1, so we would do a conSat here
- ?_3 : List ?_1, but we would not here (do a refine)