feat: Examples make (partially) saturated constructors
Note that this is a naive translation -- in particular some of the
constructors are not fully saturated (i.e. could applied to further
arguments). This will be changed in a subsequent commit. However, none
of them now appear to the left of an application node.

This is a step towards all constructors using the saturated
combinators, at which point we can remove the legacy `con`.