feat!: ConstructRefinedCon creates a saturated constructor
(In the sense of "uses `conSat`" / "embeds the arguments in the `Expr`
subtree rooted at a `Cons` node". Before this commit it created a
constructor applied to these same arguments with `APP` and `App`.)

Note that now both actions underlying the openapi `MakeConSat` action
use `conSat`.

This is a step towards all constructors being saturated, at which
point we can update the typechecker to enforce that invariant.

Note that `spanMaybe` has been resurrected from the previously-removed
primer-service/exe-replay/Main.hs

BREAKING CHANGE: this will change the interpretation of some actions
that may be stored in logs inside programs created from older primer
versions. Thus it may break usage of older databases. (The programs
should still load fine, and be editable, but undo may be broken if a
ConstructRefinedCon action happens to be in the log.)