Revert "feat: Expose undo via the Primer API & Servant endpoints."
This reverts commit d94a1b70c50dc5d3ee9765d3b41e1d63ab23bd5a.

This was an unnecessary change: undo can already be performed using
the `edit` API, as is apparent from the fact that the type signatures
for the undo operations are identical to the corresponding `edit` ops.

Arguably the descriptions for the undo-specific operations were
helpful, but I don't think they warrant the redundancy elsewhere.