feat!: remove ConstructCon, MakeCon actions as non-saturated
We leave ConstructSaturatedCon, ConstructRefinedCon and MakeConSat
which create saturated constructors.

Note that they behave the same for nullary constructors, and inserting a
non-applied constructor (which has arguments) will not make sense when
all constructors must be saturated.

This is a step towards all constructors being saturated, at which
point we can update the typechecker to enforce that invariant.

BREAKING CHANGE: this removes an action that may be stored in the undo
log of a program from a database created with an older version of
primer. This will break loading old databases.