feat!: MakeConSat nullary options and levels
Prior to this commit, "Insert a saturated constructor" did not offer
nullary constructors. Nullary constructors were offered only by "Insert
a constructor".

With this commit, "Insert a saturated constructor" now always offers
nullary constructors, and in beginner mode only offers nullary constructors.

This change is with a view to merging "Insert a saturated constructor"
and "Insert a constructor" actions.