feat: MakeConSat nullary options and levels
Prior to this commit, "Insert a saturated constructor" did not offer
nullary constructors. Nullary constructors were offered only by "Insert
a constructor".

With this commit, "Insert a saturated constructor" now always offers
nullary constructors, and in beginner mode only offers nullary constructors.

This change is with a view to merging "Insert a saturated constructor"
and "Insert a constructor" actions.

This commit changes the data we emit via the OpenAPI endpoints.
However, it obeys the same schema (and consumers are expected to not
care about the exact data we emit), and is thus not a breaking change.
Note that it does not affect the actions stored in undo logs, and thus
will not break programs loaded from older databases that were created
with older versions of primer.