feat: actions to enter/exit constructor arguments
Now that constructors directly store their arguments, we need actions
to enter (and exit) these tree positions, so one can edit them.

It would be possible to combine `Child1`, `Child2` and `ConChild`, but
for the sake of keeping this commit small and non-breaking we have added
a new type of movement.