feat!: constructors are checkable only
This has a knock-on effect on ConstructRefinedCon: if there is no
valid refinement (and we thus fall back to inserting a saturated
constructor in a hole), we need to put an annotation on the
constructor which is inside a hole.

BREAKING CHANGE: this will reject some programs that previous versions
of primer would consider valid. Thus it may break loading a database
that was generated by a older version of primer.