refactor: split generation of ADT constructor and primitive constructor/literal
This is a step towards only generating terms which are consistent with
treating constructors as checkable. That in itself is a step towards
being able to change the typechecker to treat constructors as
checkable, rather than synthesisable. This would then enable us to
drop indices from constructors.

Note that literals will be synthesisable, but constructors will only be checkable.

This commit slightly generalises what terms we generate, since we can
now generate primcons for non-hole types.