fix: TC case expression simultaneous substitution
Typechecking a 'case' expression ultimately calls 'instantiateValCons'',
which computes the types of the case-branch-bound variables, by looking
at the ADT definition and substituting in the actual values the
parameters take on. However, this was incorrectly implemented as an
iterated substitution, rather than a simultaneous substitution. Thus
we could be confused about scoping and return the wrong result, as shown
by the added test.

Since it is unused and trivial, we remove the export of the
'substituteTypeVars' helper, and inline its one use.