fix!: ConstructRefinedCon at arrow type gave unsaturated constructor
This is the last step towards all constructors being saturated. We can
now update the typechecker to enforce that invariant.

BREAKING CHANGE: this changes the interpretation of an action that may
be stored in program's undo logs. This may break the undo action of
programs loaded from a database which was created by an earlier version
of primer.