feat!: constructors do not store their indices
This changes the definition of `Con`, removing the type arguments.

This simplifies typechecking constructors (including removing some
now-unused error cases), removes the actions for entering type
children of a constructor, removes the workaround for Eval (that
focusType did not look at type arguments in constuctors) and
simplifies Eval (when doing a case reduction, we do not need to worry
about inconsistencies between indices of a constructor and its annotation).

We remove some tests that relied on having indices:
- unit_type_preservation_case_hole_regression: its starting point is impossible
- unit_con_hole_app_type_5
- unit_con_tyargs_consistent_sh

There are some test updates to pay particular attention to, since they
importantly depend on the types in constructor indices (thus not
changing the test, i.e. so these arguments are simply ignored, will
either break the test or subtly change what is being tested):
- Tests.Action.unit_rename_LAM{,_2}
- Tests.Action.Capture.unit_ty_tm_same_namespace
- Tests.Action.Prog.unit_copy_paste_expr_1
- Tests.Action.Prog.setConFieldTypeHelper
- Tests.Eval.unit_redexes_lettype_2
- Tests.Questions.unit_variablesInScope_shadowed

To avoid too much churn in this commit we make the DSL ignore these
index arguments, but still update the ID counter for them. (We will
return to this in a subsequent commit.) Some golden test output has
thus changed.

BREAKING CHANGE: this changes the definition of `Expr'`, which breaks
compatibility with databases generated by previous versions of primer.