feat!: Add redo.
We also add a few simple tests for undo and redo.

Redo is implemented as follows:

* Add a new redo log (a stack, actually) to `Prog`.

* When an undo action is successful, push the undone action onto the
redo log.

* When a redo action is requested, pop the last action off the redo
log and try to replay it. If success, return the new program state;
otherwise, throw a `ProgError` and let the caller sort it out.

Note that whenever any successful new edit is performed, we also reset
the redo log, as there would be no guarantee that actions that had
been pushed onto the redo log could be successfully replayed on the
new program state. (We may revisit this design choice in the future,
to allow for more complex undo/redo schemes.)

BREAKING CHANGE: this commit changes the `App` encoding in the
database. New programs created from this commit forward cannot be
loaded from the database by an older version of Primer. Additionally,
programs created with versions of Primer prior to this commit cannot
be loaded from the database by this version (nor going forward).
We *could* fix this at the database layer, but it's probably not worth
it at this point in Primer's evolution.