fix: SetConFieldType could create ill-typed programs
When running this action, we don't run a full typechecking pass (for
efficiency). Prior to this commit we could create ill-typed programs,
by not taking into account directionality, since arguments to
constructors may only be checkable, but the action may wrap those
terms in a hole without adding an annotation.