feat!: TypeDefs have metadata in constructor fields
This is a sprawling commit, since we change some some things deep in
the hierarchy (in particular, 'builtinModule') which have many
knock-on changes. This also means there are changes to fragile names
in the testsuite.

BREAKING CHANGE: this change requires a database migration, as it
changes the representation of 'Module'. However, since this is just
serialised to json and stored as a blob in the DB, it requires no
schema changes. Since we have no programs we need to preserve, we
decided not to bother with a migration. This means that DBs created
before this commit will not load with a primer containing this commit.