refactor: parameterise TypeDef by type metadata
We want to offer actions to edit typedefs in the OpenAPI, and thus we
want (stable) IDs in the types of constructor fields. This commit is a
small step towards that goal: we will use 'TypeDef ()' everywhere for
now. In a subsequent commit we will use 'TypeDef ()' where we don't
care about this metadata (e.g. the typechecker), and 'TypeDef
TypeMeta' where we do.