refactor: further simplify typedef actions now constructors are saturated
In the handlers for `SetConFieldType` and `AddConField` we can now use
`transformM` instead of manually doing the recursion.

This change was enabled by the previous commit, which removed the
handling of non-saturated constructors. Prior to that we had to recurse
top-down, since for an unsaturated term such as `C x xs` (which would be
represented as `App (App (Con "C" [] []) (Var "x")) (Var "xs")) we need
to act on the outer `App` node and then recurse into the arguments
(here, `x` and `xs`). Note that we must not act on the "inner" `App`
node, since we would then believe that the constructor is only applied
to one argument. Thus `transformM` is inappropriate as it is a bottom-up
traversal, and we must hand-roll the traversal with `descendM`.

Since we now know that all constructors are saturated, we know that
the above situation is imposible, and the representation must in fact be
`Con "C" [] [Var "x", Var "xs"]`. This sidesteps all the issues,
enabling us to use the standard bottom-up traversal `transformM`.