refactor: mapOddProg ID ordering
This changes the order of operations in a `do` block, which will have
the effect of changing what IDs each node gets assigned. The reason
for this is somewhat technical and somewhat trivial: it happens to be
necessary so we can split some future changes up into separate
commits. The particular change we are aiming for is to drop type
arguments from constructors. We'll reduce churn in the main commit by
changing the DSL's `con` to ignore its type argument but still tick
the fresh ID counter for it (thus we can make the change to the AST
without updating much of the testsuite). There will then be a followup
which mechanically drops all the ignored arguments. There is a small
wrinkle in this plan: if `con` is used in such a way as to ignore a
node which would previously be assigned the highest ID in a program,
then a test that checks that the `nextProgID` of that program is what
was returned by `create` will fail. This actually happens in exactly
one place: `Tests.App.unit_mkAppSafe_mapOddProg`. We solve this by a
simple (and subtle) reordering so the last ID in `mapOddProg` is in
its type, and thus not dropped.