feat!: internals of non-empty holes check against type hole
Previously the typing rule was
   e ∈ S
-----------
{? e ?} ∈ ?

Now the rule is
   ? ∋ e
-----------
{? e ?} ∈ ?

BREAKING CHANGE: since this changes the behaviour of smartholes and the
`SetConFieldType` action, it will change the interpretation of program
logs. Thus it may break replaying logs stored from older versions of
primer (e.g. when using 'undo' on a program loaded from an old
database).

Signed-off-by: Ben Price <ben@hackworthltd.com>