fix: offered names for renaming `let`s
Previously the names were based on the type of the let itself, rather
than the type of the bound expression.

We add a test to the testsuite, and to do so we change
`offeredActionTest` to typecheck the `Prog` to fill in the
typecaches (which is what drives the offered names).

Signed-off-by: Ben Price <ben@hackworthltd.com>