fix: offered names for renaming a forall
Note that this does not affect the output of any golden tests. This is
because the only tests that could be affected are the outputs of
available actions for `comprehensive`. This happens to only contain
foralls where the bound variable is of kind `KType`. Such foralls are
unaffected as bodies of foralls are always of kind `KType` and the bug
fixed in this commit was confusing the kind of the body with the kind of
the variable (both `KType` in these instances).

Signed-off-by: Ben Price <ben@hackworthltd.com>