refactor!: remove redundant ConstructRefinedCon
Now that constructors are fully saturated and do not store their
indices, there is essentially no difference between inserting a
saturated constructor and a refined constructor. This is because they
both insert that constructor applied to the same number empty term
holes (dictated by the number of fields that constructor has). In the
past, when constructors were applied to types corresponding to the
parameters of their parent datatype definition, "saturated" would put
empty type holes, whereas "refine" would infer what types it must be
applied to from context.

Note that above we said "essentially no difference". The inessential
differences occur when one of the following is true
- there are internal failures in refinement, which should never happen
- smartholes are off, which is an unsupported mode for clients (it is
  intended for internal testing purposes)
- one tries to "refine" in a synthesisable-only context, which shouldn't
  happen with an openapi client, since we wouldn't offer that action in
  that context anyway

BREAKING CHANGE: this removes an action that may have been stored in an
undo log in a database created by a previous version of primer, thus
breaking loading such DBs.