chore: Add a DCO check.
Closes https://github.com/hackworthltd/primer/issues/936

Note that this workflow currently triggers on merge queue entries, but
I'm not sure whether that functionality will work, because the Probot
app upon which it's based doesn't have explicit support for that.
However, it may not matter in practice, as presumably any PR that
enters the merge queue will have already passed the DCO check in the
first place.

For now, this check isn't a required job, only advisory, so it won't
block anything from merging, in any case.

Signed-off-by: Drew Hess <src@drewhess.com>