doc: add a `CONTRIBUTING.md` document
Closes https://github.com/hackworthltd/primer/issues/15

Note that this doesn't yet address our commit message requirements,
because I need to spend a bit of time fiddling with `commitlint` and
friends, first. I'll update the document when I've had a chance to
lock all of that down.

The document also describes some buttons and forms in this repo's
GitHub Issues that don't quite match up with our current
configuration. I'll address that as well in subsequent commits.

Signed-off-by: Drew Hess <src@drewhess.com>