doc: Add a project overview, a project FAQ, and a table of contents.
Note: both the project overview and the FAQ are intended to
cover *this implementation* of the Primer programming language, not
the language itself. Think of these as documents for people who want
to understand how the implementation works at a high level, how the
various Haskell libraries fit together, and why we made the
implementation choices we did.

Signed-off-by: Drew Hess <src@drewhess.com>