feat: find sessions by substring match on session name
Note that this particular substring search matches anywhere within a
session name, and is case-sensitive.

Because our pagination API reports `n` results from `m` total, we need
the total query count. This initial implementation uses a particularly
inefficient method of counting the total number of matches (it runs
the query twice). I think this should be straightforward to fix, but
might require some refactoring. This inefficiency will be fixed in
subsequent work.

Signed-off-by: Drew Hess <src@drewhess.com>