feat: add `findSessions` to the HTTP APIs
Note that we do this not by adding a `findSessions` API endpoint
directly, but by merging the functionality into the existing
`getSessionList` API endpoints. This is in keeping with REST-ish API
conventions, where a resource endpoint (the list of sessions, in this
case) can be filtered based on query parameters, as opposed to 2
separate "list all" and "filter all" endpoints.

(We do not take this approach in the Haskell API, also for reasons of
conventionality.)

Signed-off-by: Drew Hess <src@drewhess.com>