feat!: remove "only in-memory" option for `listSessions`
We never used this "feature," which was only ever intended for testing
the `flushSessions` API, anyway. Removing the option means we don't
need to add a matching option to `findSessions` in order to fuse it,
in the OpenAPI API, with the `listSessions` API call, which is
something we'll need to do in an upcoming commit to support REST API
conventions.

We could, instead, simply not expose the parameter from the HTTP APIs,
but as we're not using it, even for testing, we might as well just
simplify everything and take it out altogether.

BREAKING CHANGE: this removes a parameter from the `listSessions` API
call and its HTTP API equivalents.

Signed-off-by: Drew Hess <src@drewhess.com>