feat: add a paranoid guard against SQLite SQL injection attacks
Because we can't properly escape search substrings using Selda, we
take a cautious approach and replace occurrences of `%` in the search
string with `_`, which will only match one character, as opposed to
`%`, which is eager. We'll still find names with literal `%` in them,
in case the intentions were genuine, we'll just return more than we
should.

Signed-off-by: Drew Hess <src@drewhess.com>