chore: change occurrences of "user" -> "student"
Per the terminology we've been using since the project's inception.
We've done a good job with this overall, we've just missed a
few. (Note that there are still a few occurrences of "user" in the
code base and project documentation, but those uses don't refer to
people who're using Primer, which is the use case we care about.)

Signed-off-by: Drew Hess <src@drewhess.com>