refactor!: introduce a notion of a Pattern
As setup for enabling pattern matching on primitives, we introduce a
notion of a `Pattern`. This is currently just a `ValConName` (since all
pattern matches are specified up to alpha equivalence by their head
constructor) , but later will have another option to be a `PrimCon`
naming branches on a primitive type.

BREAKING CHANGE: this changes actions, which are stored in programs
logs. These are visible via the strongly-typed api and in saved databases.

Signed-off-by: Ben Price <ben@hackworthltd.com>