feat!: high-level actions for case on primitives
Note that there is no way in our API to say "this action requires an
argument, and the argument can be any Int except 3 or 7". Thus when
adding a branch to a pattern match of primitive type, we cannot request
a frontend to give a branch that does not currently exist. We will give
an error if an attempt to add a currently-existing branch is made.

It would be possible to augment the API's notion of `FreeInt` and
`FreeChar` to include a finite list of illegal entries, but we leave
this as a possibility for future work.

BREAKING CHANGE: this expands the OpenAPI

Signed-off-by: Ben Price <ben@hackworthltd.com>