refactor!: branch moves
As setup for enabling moving to a fallback branch, we introduce a notion
of a `BranchMove`. This is currently just the `ValConName` of the branch
to move to, but later will have another option to move to the fallback
branch.

BREAKING CHANGE: this changes actions, which are stored in programs
logs. These are visible via the strongly-typed api and in saved databases.

Signed-off-by: Ben Price <ben@hackworthltd.com>