feat!: Case nodes now can have a wildcard/fallback branch
This means that a pattern match can explicitly match on a subset of the
constructors of a type, and give a final "catch-all" branch to match any
other constructor. This does not give any extra expressivity, since it
is possible to exhaustively list every branch, duplicating the
output (since all ADTs are finitely branching). However, this is setup
to enable pattern matching on primitives in future work, since
primitives are infinitely (or impractically large, but finitely)
branching.

Note that we expose a new DSL combinator `caseFB_` instead of modifying
the existing `case_` to avoid churn and the noise of having
`CaseExhaustive` in many places.

BREAKING CHANGE: this changes the AST of core primer terms, affecting
programs stored in a database and the strongly-typed API. This also adds
some new flavors to the OpenAPI.

Signed-off-by: Ben Price <ben@hackworthltd.com>