feat: trivial pattern matches are lazy
Consider a pattern match which has only a falllback branch:
`case e of _ -> t`; before this commit, evaluating this would force `e`
to WHNF before reducing the `case`; after this commit this will reduce
to `t` even if `e` diverges. Note that a case with at least one explict
match must still force `e`, so one can recover the old behaviour by
writing `case e of C x y z -> t ; _ -> t`, duplicating the `t`; however,
this is only possible for non-empty types.

We make this design decision so that when a future extension enables
deep pattern matching the semantics are more obvious and
consistent (since a wildcard is then the same as a pattern variable
which happens not to be referenced).

Signed-off-by: Ben Price <ben@hackworthltd.com>