fix: use cached type of scrut in transformCaseBranches
Instead of synthesising the type of the scrutinee from scratch, we just
read it from the TypeCache. This relies on the program always being in a
consistent state, which is tested by tasty_available_actions_accepted,
at least for interactions using SmartHoles and the "offered action" api.

This also fixes a bug: previously when we synthesised the type, we
needed to know the context that the scrutinee lives in. However, we only
used the `progCxt`, which is the "top-level"/"global" context -- in
particular this would fail if the scrutinee used a variable introduced
by a lambda (or let, etc).

Signed-off-by: Ben Price <ben@hackworthltd.com>