feat: optionally do not evaluate under binders
This change means our EvalFull is "closed evaluation", rather than
"open evaluation". The main reason to do this is so that evaluations of
programs-in-construction (especially recursive ones) do not pointlessly
blow up in size: if `even` and `odd` are defined recursively, then
evaluating `even`  would evaluate under the lambda and inside case
branches, expanding `odd` and recursing; when it eventually times out
one would have a big tree with many unrolled copies of `even` and `odd`,
which is not very illuminating.

Note that the reduction of "pushing a let down", e.g.
`let x = t in f s  ~> (let x = t in f) (let x = t in s)` happens at the
`let` node, and thus is not "under" the `let x` binder.

Note that we consider a "stack" of lets `let x=s in let y=t in r` to be
one binder, so that we can (with `groupedLets = False`) reduce this by
pushing the `let y` into the `r` (i.e. we don't count this as "under"
the `let x` binder). We will not evaluate inside the `r`.

Note that we do not evaluate in the RHS of pattern match branches which
bind variables, for the same reason as we do not evaluate under lambdas;
for consistency we also do not evaluate in the RHS of branches which do
not bind variables.

Signed-off-by: Ben Price <ben@hackworthltd.com>