fix: forgetProgTypecache also acts on typedefs
Since 6197cc180dfaf7c7ba18e0a8352da684516e64d7, type definitions have
contained metadata (since they need IDs so that we can run actions on
them). When forgetting metadata in a Prog, we also need to forget this
metadata.

The reason we have noticed this now is that we are aiming to upgrade the
`tasty_available_actions_accepted` test to cover typedef actions which
will require upgrading `checkEverything` to modify typedefs in a
subsequent commit (it may need to do smartholes, rather than just saying
"no"). As it will then update typecaches, the benchmark tests would
start to fail as (before this current commit) we were not properly
forgetting them.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>