fix: ignore typedef name clashes for available-actions-accepted
This wasn't actually due to test failures, just something I noticed in
passing. The generators which create the initial program are very
unlikely to pick the same names as the test will try to create -- we
leave tweaking the generators to try to exercise this sort of thing more
as future work. (Also note that adding a typedef is not an "offered"
action, so that error wouldn't be hit here, however there is no harm in
explicitly ignoring it.)

Signed-off-by: George Thomas <georgefsthomas@gmail.com>
Signed-off-by: Ben Price <ben@hackworthltd.com>