refactor: export `allNonPrimValCons` convenience function
This function was previously embedded in `lookupConstructor`, but is
useful elsewhere. We use it in this commit to DRY a bit in the offered
options for `MakeCon`. (We don't use the extra `TyConName` tuple
parameter it provides yet, but we will in the next commit.)

Note that there are no functional changes here; this is purely a
refactor.

Signed-off-by: Drew Hess <src@drewhess.com>