feat: annotate offered valcons that match the selection's type
This is an extension of
https://github.com/hackworthltd/primer/pull/1045, which itself was a
proof of concept/best effort to annoatate variables that match the
type of a hole up to alpha. Here we extend this feature to
non-primitive value constructors.

This implementation uses `valConType` to extract the value
constructor's type. For the record, the documentation for `conInfo`,
which also uses `valConType`, reads as follows:

"the type an eta-expanded version of this constructor would check
against e.g. `Cons`'s "type" is `∀a. a -> List a -> List a`."

Signed-off-by: Drew Hess <src@drewhess.com>