fix!: typedef actions detects all name clashes
Previously we did not consistently check our input names were fresh,
leading to some actions being accepted, only to later throw a
typechecker error. We now always check the new name is fresh, but for
ease of maintainability we collapse all such errors into one
category (previously we had both `ParamAlreadyExists` and
`TyConParamClash`). Instead of collapsing them, we could split out 5
different sorts of clashes (see comment on `TypeDefModifyNameClash` in
this commit), but this does not seem necessary -- as far as I know,
nobody actually wants to distinguish which of these 5 sorts of clashes
happened.

BREAKING CHANGE: this changes `ProgError`, which is serialised in the
richly-typed API (but not the OpenAPI -- they are converted into http
error codes instead).

Signed-off-by: Ben Price <ben@hackworthltd.com>