refactor: explicit TC error for duplicate parameters
We now explicitly distinguish "two parameters in a typedef are named the
same" and "a parameter is named the same as a constructor, or two
constructors are named the same". (This is a somewhat silly split, as
the second one arguably would be better as two separate checks. The
reason it is broken up this way is that the first makes sense for
primitive typedefs, and the second does not.)

Previously, for ADTs, a repeated parameter name would have been caught
by the check that all parameters+ctors are distinctly named. However, we
now call `checkTypeDef` on primitives as well, which would not be
covered by the ADT check (but primitives are hardcoded and don't repeat
parameter names, so this is a somewhat pointless check).

Signed-off-by: Ben Price <ben@hackworthltd.com>