refactor: AppTestM's MonadFresh instance uses App
Previously `AppTestM` transformed `TestM`, thus it had both an `App`
state (recall, an `App` contains some fresh `ID` and `NameCounter`
state) and also a separate `ID`/`NameCounter` state via `TestM`. Thus,
some things use one set of counters, and some the other. In particular,
`handleRedoRequest` works via `runEditAppM` which uses the `App`'s
counters, but `handleEditRequest` uses the `TestM`'s counters.

This only happened in tests, and not in the actual service. This commit
ensures that tests are closer to the real service by removing the
`TestM` layer, ensuring that we always use the `App`'s counter. (This is
needed to tickle some long-standing bugs involving "lost IDs" in the
undo/redo system.)

Signed-off-by: Ben Price <ben@hackworthltd.com>